import/python/requirements.txt
neo4j-driver==1.4.0
